Added Refactored decoupled components

+++ include/Keypad.hpp
@@ -3,6 +3,7 @@
 
 #include <Constants.hpp>
 #include <array>
+#include <cstdint>
 
 namespace chip8::core {
 class Keypad {
@@ -13,7 +14,7 @@ class Keypad {
 
     bool set_state(int key_idx, bool state);
     bool is_pressed(int key_idx) const;
-    int get_pressed() const;
+    uint8_t get_pressed() const;
     void reset();
 };
 } // namespace chip8::core
